Updated versions for Appcenter and app name int titlebar With regard to removing a warning in AppCenter build, adding this to the gradle file.  lintOptions {         checkReleaseBuilds false     } The warning is "mylibrary1 is either no Android app project or build version has not been set to override. Skipping..." Check to see if it goes away. (https://stackoverflow.com/questions/62556363/appcenter-build-error-project-firebaseauth-web-at-is-either-no-android-app-pr)
With regard to networkinterface privacy and not able to get interface wlan0, adding permissions in library3
(fun_getMACAddress.kt) in manifest although it did not fix the problem. Lots of sample code uses it though.
<uses-permission android:name="android.permission.INTERNET" />
<uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />

Updated strings.xml and build.gradle for version and to kick to do autoupdate in app
Updated to androidx.constraintlayout:constraintlayout:2.0.4
More fixes to get AppCenter to recognize the supposed to be signed build
Added things for signing. Appcenter can't distribute apps that aren't signed
Rename app_name to PocketKnife Proof of Buildout 2

Do this each time and update the versionCode to force a release from AppCenter
Ready to try in-app updates. Upped the versionCode to 2
Moved image folder to the root from app
Images for readme.md
Edited readme.md
Prepping for Appcenter

Finalizedfunr_getMACAddress code to handle network interface privavacy change starting in Android 28 or so.
REMOVED this code:
Added code to work around Android 28 (29?) and above not giving up MAC addr for privacy
    if (checkOSversion())
        return builder.append(
            mActivity.getString(R.string.mac_address_notsupported)
        ).toString()

Updating to make better
inital commit